% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_EMR_to_cohort.R
\name{better_calendar}
\alias{better_calendar}
\title{Function better_calendar()}
\usage{
better_calendar(totranslate)
}
\arguments{
\item{totranslate}{document with dates in letters (class : character)}
}
\value{
totranslate : formated text (class : character)
}
\description{
This function allows you to transform all dates that are in letters into a standardized format DD/MM/YYYY.
}
\examples{
package needed for the example: library(data.table)
res.lcm <- res.lcm[,TEXTE_AFFICHAGE := lapply(TEXTE_AFFICHAGE, function(row) better_calendar(totranslate = row))]
res.lcm[["TEXTE_AFFICHAGE"]] <- as.character(res.lcm[["TEXTE_AFFICHAGE"]])
better_calendar()
}
\keyword{calendar}
\keyword{datamart}
\keyword{date}
