% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_EMR_to_cohort.R
\name{blindeye}
\alias{blindeye}
\title{Function blindeye()}
\usage{
blindeye(df_to_anon, algo, size_key)
}
\arguments{
\item{df_to_anon}{datamart that will be de-identified (class : data.frame). Columns needed : TEXTE_AFFICHAGE (doc in character class), NOM (name), PRENOM (surname), DATENAIS(birthday date).}

\item{algo}{name of the hashing algorithm chosen from the "openssl" package, "sha512" or "sha384" are advised because they are more secure (sha-2 family). (Class : character)}

\item{size_key}{randomly generated hash key, a size of at least 64 is advised (class : integer)}
}
\value{
list_return :

list_return$df : the de-identified datamart (class : data.frame)

list_return$liste : liste of all the patients and their identifying data after de-identification of the datamart(class : data.frame)(need to be kept secret).

list_return$hashkey : the key used to hash the specific hospital code of each patient (IPP)(class : character)(need to be kept secret).
}
\description{
This function allows you to de-identify the patients in the text from Name, surname, date of birth (replaced by age), Id_code in the hospital (IPP). It also keep everything in order to re-identify the patients if needed.
}
\examples{
returnblind <- blindeye(df_to_anon = res.lcm, algo = "sha512", size_key = 64)
blindeye()
}
\keyword{datamart}
\keyword{de-identification}
\keyword{hash}
