% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_EMR_to_cohort.R
\name{getting_sight_back}
\alias{getting_sight_back}
\title{Function getting_sight_back()}
\usage{
getting_sight_back(df_to_anon, algo, liste_anon, hashkey)
}
\arguments{
\item{WORKINPROGRESS}{WORKINPROGRESS}
}
\value{
WORKINPROGRESS
}
\description{
This function allows you to re-identify the datamart (in pair with the blindeye() function)
}
\examples{
WORKINPROGRESS
getting_sight_back()
}
\keyword{datamart}
\keyword{re-identification}
