% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_EMR_to_cohort.R
\name{html_xmltotext}
\alias{html_xmltotext}
\title{Function html_xmltotext()}
\usage{
html_xmltotext(rowdoc)
}
\arguments{
\item{rowdoc}{document that needs removal of its tags (class : character)}
}
\value{
doctxt : document without html or xml tags, ready for extraction if it is a table (class : character).
}
\description{
This function allows you to remove all html and xml tags from the documents, but keep the structure of the document.
}
\examples{
package needed for the example: library(data.table)
res.lcm <- res.lcm[,TEXTE_AFFICHAGE := lapply(TEXTE_AFFICHAGE, function(row) html_xmltotext(rowdoc = row))]
html_xmltotext()
}
\keyword{datamart}
\keyword{html}
\keyword{xml}
