% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_EMR_to_cohort.R
\name{offset_date_text}
\alias{offset_date_text}
\title{Function offset_date_text()}
\usage{
offset_date_text(textpurified, offset_date)
}
\arguments{
\item{textpurified}{document which dates will be modified, purified from dates in letters (class : character).}

\item{offset_date}{number of weeks of the offset (class : integer).}
}
\value{
textpurified : text offsetted (class : character)
}
\description{
This function allows you to change(offset) all dates in standardized format DD/MM/YYYY, by a number of weeks.
}
\examples{
res.lcm[["TEXTE_AFFICHAGE"]] <- apply(X = res.lcm, MARGIN = 1, function(row_dt) offset_date_text(textpurified = row_dt[["TEXTE_AFFICHAGE"]], offset_date = row_dt[["WEEK_OFFSET"]]))
offset_date_text()
}
\keyword{datamart}
\keyword{date}
\keyword{offset}
